+++ applications/fishing-map/src/features/map/map-viewport.hooks.ts
@@ -15,7 +15,7 @@ import { selectViewport } from 'features/app/app.selectors'
 import store, { RootState } from '../../store'
 import useMapInstance from './map-context.hooks'
 
-type SetMapCoordinatesArgs = Pick<ViewportProps, 'latitude' | 'longitude' | 'zoom'>
+type SetMapCoordinatesArgs = Partial<Pick<ViewportProps, 'latitude' | 'longitude' | 'zoom'>>
 type UseViewport = {
   viewport: MapCoordinates
   onViewportChange: (viewport: ViewportProps) => void
@@ -52,8 +52,8 @@ const viewportState = atom<MapCoordinates>({
 export default function useViewport(): UseViewport {
   const [viewport, setViewport] = useRecoilState(viewportState)
 
-  const setMapCoordinates = useCallback((viewport: SetMapCoordinatesArgs) => {
-    setViewport({ ...viewport })
+  const setMapCoordinates = useCallback((coordinates: SetMapCoordinatesArgs) => {
+    setViewport((viewport) => ({ ...viewport, ...coordinates }))
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])