+++ packages/layer-composer/src/generators/track/track.ts
@@ -9,7 +9,7 @@ import { Group } from '../../types'
 import { Type, TrackGeneratorConfig, MergedGeneratorConfig } from '../types'
 import { memoizeByLayerId, memoizeCache } from '../../utils'
 import { isConfigVisible } from '../utils'
-import filterGeoJSONByTimerange from './filterGeoJSONByTimerange'
+import filterTrackByTimerange from './filterTrackByTimerange'
 import { simplifyTrack } from './simplify-track'
 
 const mapZoomToMinPosΔ = (zoomLoadLevel: number) => {
@@ -42,7 +42,7 @@ const filterByTimerange = (data: FeatureCollection, start: string, end: string) 
   const startMs = new Date(start).getTime()
   const endMs = new Date(end).getTime()
 
-  const filteredData = filterGeoJSONByTimerange(data, startMs, endMs)
+  const filteredData = filterTrackByTimerange(data, startMs, endMs)
   return filteredData
 }
 
@@ -54,7 +54,7 @@ const getHighlightedData = (
   const startMs = new Date(highlightedStart).getTime()
   const endMs = new Date(highlightedEnd).getTime()
 
-  const filteredData = filterGeoJSONByTimerange(data, startMs, endMs)
+  const filteredData = filterTrackByTimerange(data, startMs, endMs)
 
   return filteredData
 }