+++ applications/tile-inspector/CHANGELOG.md
@@ -1,9 +1,22 @@
 # tile-inspector
 
+## 0.1.47
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@8.1.0
+  - @globalfishingwatch/react-hooks@11.1.0
+  - @globalfishingwatch/ui-components@8.1.0
+
 ## 0.1.46
 
 ### Patch Changes
 
+- @globalfishingwatch/api-client@2.0.0
+- @globalfishingwatch/layer-composer@8.0.0
+- @globalfishingwatch/react-hooks@11.0.0
+- @globalfishingwatch/ui-components@8.0.0
 - Updated dependencies [fdeabdfa]
   - @globalfishingwatch/ui-components@6.0.1
 
@@ -11,6 +24,10 @@
 
 ### Patch Changes
 
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@7.3.0
+  - @globalfishingwatch/react-hooks@10.0.0
+  - @globalfishingwatch/ui-components@7.0.0
 - Updated dependencies [05fe3de6]
   - @globalfishingwatch/layer-composer@7.2.0
   - @globalfishingwatch/react-hooks@9.0.0
@@ -20,6 +37,9 @@
 
 ### Patch Changes
 
+- Updated dependencies [undefined]
+  - @globalfishingwatch/react-hooks@9.0.1
+  - # @globalfishingwatch/ui-components@6.0.1
 - Updated dependencies [f96a081c]
   - @globalfishingwatch/ui-components@5.1.2
 
@@ -28,6 +48,9 @@
 ### Patch Changes
 
 - Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@7.2.0
+  - @globalfishingwatch/react-hooks@9.0.0
+  - @globalfishingwatch/ui-components@6.0.0
   - @globalfishingwatch/ui-components@5.1.1
 
 ## 0.1.42