+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @globalfishingwatch/react-hooks
 
+## 11.1.0
+
+### Minor Changes
+
+- fishing events
+
+## 11.0.0
+
+### Patch Changes
+
+- @globalfishingwatch/api-client@2.0.0
+- @globalfishingwatch/dataviews-client@11.0.0
+- @globalfishingwatch/layer-composer@8.0.0
+
+## 10.0.0
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@7.3.0
+  - @globalfishingwatch/dataviews-client@10.0.0
+
+## 9.0.1
+
+### Patch Changes
+
+- events generator
+- Updated dependencies [undefined]
+  - @globalfishingwatch/dataviews-client@9.0.1
+
+## 9.0.0
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/dataviews-client@9.0.0
+  - @globalfishingwatch/layer-composer@7.2.0
+
 ## 9.0.0
 
 ### Major Changes