+++ packages/timebar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/timebar
 
+## 1.8.0
+
+### Minor Changes
+
+- fishing events
+
+## 1.7.0
+
+### Minor Changes
+
+- currentEvent interaction
+
 ## 1.6.1
 
 ### Patch Changes